Split `WebiServer` from `WebiOutput`.
