+++ crate/webi_output/src/lib.rs
@@ -1,7 +1,8 @@
 //! Web interface output for the peace automation framework.
 
-pub use crate::webi_output::WebiOutput;
+pub use crate::{webi_output::WebiOutput, webi_server::WebiServer};
 
 pub mod assets;
 
 mod webi_output;
+mod webi_server;
